Add movie card layout Add movie card layout with placeholder info

+++ app/src/main/java/com/suslanium/filmus/domain/entity/movie/MovieSummary.kt
@@ -10,5 +10,5 @@ data class MovieSummary(
     val country: String?,
     val genres: List<Genre>,
     val rating: Float?,
-    val userRating: Float?
+    val userRating: Int?
 )
